backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
requests==2.31.0
openai>=1.52.0
mcp>=1.0.0
google-genai>=1.0.0